Tests/LibWeb/Text/expected/trusted-event.txt

beforetoggle event (ToggleEvent) is trusted
beforetoggle event (ToggleEvent) is trusted
toggle event (ToggleEvent) is trusted
toggle event (ToggleEvent) is trusted
beforeunload event (BeforeUnloadEvent) is trusted
unhandledrejection event (PromiseRejectionEvent) is trusted
